Kotlin files to collections package and remove unwanted strings in strings file
