+++ app/src/main/AndroidManifest.xml
@@ -13,19 +13,19 @@
         android:theme="@style/Theme.JOM.Light"
         tools:targetApi="31">
         <activity
-            android:name=".CollectionCompletedActivity"
+            android:name=".collections.CollectionCompletedActivity"
             android:exported="false" />
         <activity
-            android:name=".OptionalVerificationActivity"
+            android:name=".collections.OptionalVerificationActivity"
             android:exported="false" />
         <activity
-            android:name=".VerifyAmountActivity"
+            android:name=".collections.VerifyAmountActivity"
             android:exported="false" />
         <activity
-            android:name=".CompleteCollectionActivity"
+            android:name=".collections.CompleteCollectionActivity"
             android:exported="false" />
         <activity
-            android:name=".ViewCollectionActivity"
+            android:name=".collections.ViewCollectionActivity"
             android:exported="true">
         </activity>
         <activity
