+++ app/src/main/java/org/jom/collector/collections/OptionalVerificationActivity.kt
@@ -1,4 +1,4 @@
-package org.jom.collector
+package org.jom.collector.collections
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
@@ -7,15 +7,17 @@ import android.widget.Button
 import android.widget.ImageView
 import androidx.core.content.ContextCompat
 import com.google.android.material.bottomnavigation.BottomNavigationView
+import org.jom.collector.DashboardActivity
+import org.jom.collector.R
 
-class CollectionCompletedActivity : AppCompatActivity() {
+class OptionalVerificationActivity : AppCompatActivity() {
 
     private lateinit var bottomNavigationView: BottomNavigationView
     private lateinit var backButton: ImageView
-    private lateinit var dashboard: Button
+    private lateinit var verify: Button
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_collection_completed)
+        setContentView(R.layout.activity_optional_verification)
 
         //back
         backButton = findViewById(R.id.back_button)
@@ -25,13 +27,12 @@ class CollectionCompletedActivity : AppCompatActivity() {
         window.navigationBarColor = ContextCompat.getColor(this, R.color.lightPrimaryColor)
         window.statusBarColor = ContextCompat.getColor(this, R.color.lightPrimaryColor)
 
-        // optional verification
-        dashboard = findViewById(R.id.dashboard)
-        dashboard.setOnClickListener {
-            val intent = Intent(this, DashboardActivity::class.java)
+        // complete collection
+        verify = findViewById(R.id.verify)
+        verify.setOnClickListener {
+            val intent = Intent(this, CollectionCompletedActivity::class.java)
             startActivity(intent)
         }
-
 
         // bottom nav handler
         bottomNavigationView = findViewById(R.id.bottom_nav)
